% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scip_version.R
\name{scip_version}
\alias{scip_version}
\title{\emph{SCIP} version}
\usage{
scip_version()
}
\value{
A \code{character} value containing the version number.
}
\description{
Get the version number of the \emph{SCIP} software installed.
}
\examples{
# get version number
scip_version()

}
